Refactor user detail view to dedicated page
