--- components/users/user-detail-dialog.tsx
@@ -1,682 +1,0 @@
-"use client"
-
-import { useEffect, useState } from "react"
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog"
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
-import { Badge } from "@/components/ui/badge"
-import { Button } from "@/components/ui/button"
-import { Card } from "@/components/ui/card"
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog"
-import {
-  Mail,
-  MailCheck,
-  Trophy,
-  BookOpen,
-  HelpCircle,
-  TrendingUp,
-  Shield,
-  Lock,
-  Unlock,
-  Trash2,
-  RotateCcw,
-  KeyRound,
-  Building2,
-  Calendar,
-  Globe,
-} from "lucide-react"
-import type { User } from "@/types/user"
-import type { Enrollment, QuizAttempt } from "@/types/common"
-import { api } from "@/lib/api/exports"
-import { useToast } from "@/hooks/use-toast"
-
-interface UserDetailDialogProps {
-  open: boolean
-  onClose: () => void
-  user: User | null
-}
-
-export function UserDetailDialog({ open, onClose, user: initialUser }: UserDetailDialogProps) {
-  const [user, setUser] = useState<User | null>(initialUser)
-  const [enrollments, setEnrollments] = useState<Enrollment[]>([])
-  const [quizAttempts, setQuizAttempts] = useState<QuizAttempt[]>([])
-  const [isUpdating, setIsUpdating] = useState(false)
-  const [showLockDialog, setShowLockDialog] = useState(false)
-  const [showDeleteDialog, setShowDeleteDialog] = useState(false)
-  const [showResetPasswordDialog, setShowResetPasswordDialog] = useState(false)
-  const [lockoutDays, setLockoutDays] = useState("7")
-  const { toast } = useToast()
-
-  useEffect(() => {
-    setUser(initialUser)
-    if (initialUser) {
-      loadUserData()
-    }
-  }, [initialUser])
-
-  const loadUserData = async () => {
-    if (!initialUser) return
-    try {
-      const [enrollmentsData, attemptsData] = await Promise.all([
-        api.enrollments.getByUserId(initialUser.id),
-        api.quizAttempts.getByUserId(initialUser.id),
-      ])
-      setEnrollments(enrollmentsData)
-      setQuizAttempts(attemptsData)
-    } catch (error) {
-      console.error("[v0] Failed to load user data:", error)
-    }
-  }
-
-  const handleRoleChange = async (newRole: User["role"]) => {
-    if (!user) return
-    setIsUpdating(true)
-    try {
-      const updatedUser = await api.users.updateRole(user.id, newRole)
-      setUser(updatedUser)
-      toast({
-        title: "Role updated",
-        description: `User role changed to ${newRole}`,
-      })
-    } catch (error) {
-      toast({
-        title: "Error",
-        description: "Failed to update user role",
-        variant: "destructive",
-      })
-    } finally {
-      setIsUpdating(false)
-    }
-  }
-
-  const handleTenantChange = async (newTenant: string) => {
-    if (!user) return
-    setIsUpdating(true)
-    try {
-      const updatedUser = await api.users.updateTenant(user.id, newTenant)
-      setUser(updatedUser)
-      toast({
-        title: "Tenant updated",
-        description: `User tenant changed to ${newTenant}`,
-      })
-    } catch (error) {
-      toast({
-        title: "Error",
-        description: "Failed to update user tenant",
-        variant: "destructive",
-      })
-    } finally {
-      setIsUpdating(false)
-    }
-  }
-
-  const handleLockAccount = async () => {
-    if (!user) return
-    setIsUpdating(true)
-    try {
-      const lockoutUntil = new Date()
-      lockoutUntil.setDate(lockoutUntil.getDate() + Number.parseInt(lockoutDays))
-      const updatedUser = await api.users.lockAccount(user.id, lockoutUntil.toISOString())
-      setUser(updatedUser)
-      setShowLockDialog(false)
-      toast({
-        title: "Account locked",
-        description: `Account locked until ${lockoutUntil.toLocaleDateString()}`,
-      })
-    } catch (error) {
-      toast({
-        title: "Error",
-        description: "Failed to lock account",
-        variant: "destructive",
-      })
-    } finally {
-      setIsUpdating(false)
-    }
-  }
-
-  const handleUnlockAccount = async () => {
-    if (!user) return
-    setIsUpdating(true)
-    try {
-      const updatedUser = await api.users.unlockAccount(user.id)
-      setUser(updatedUser)
-      toast({
-        title: "Account unlocked",
-        description: "User can now access their account",
-      })
-    } catch (error) {
-      toast({
-        title: "Error",
-        description: "Failed to unlock account",
-        variant: "destructive",
-      })
-    } finally {
-      setIsUpdating(false)
-    }
-  }
-
-  const handleSoftDelete = async () => {
-    if (!user) return
-    setIsUpdating(true)
-    try {
-      const updatedUser = await api.users.softDelete(user.id)
-      setUser(updatedUser)
-      setShowDeleteDialog(false)
-      toast({
-        title: "Account deleted",
-        description: "User account has been soft deleted",
-      })
-    } catch (error) {
-      toast({
-        title: "Error",
-        description: "Failed to delete account",
-        variant: "destructive",
-      })
-    } finally {
-      setIsUpdating(false)
-    }
-  }
-
-  const handleRestore = async () => {
-    if (!user) return
-    setIsUpdating(true)
-    try {
-      const updatedUser = await api.users.restore(user.id)
-      setUser(updatedUser)
-      toast({
-        title: "Account restored",
-        description: "User account has been restored",
-      })
-    } catch (error) {
-      toast({
-        title: "Error",
-        description: "Failed to restore account",
-        variant: "destructive",
-      })
-    } finally {
-      setIsUpdating(false)
-    }
-  }
-
-  const handleResetPassword = async () => {
-    if (!user) return
-    setIsUpdating(true)
-    try {
-      const result = await api.users.resetPassword(user.id)
-      setShowResetPasswordDialog(false)
-      toast({
-        title: "Password reset sent",
-        description: result.message,
-      })
-    } catch (error) {
-      toast({
-        title: "Error",
-        description: "Failed to send password reset",
-        variant: "destructive",
-      })
-    } finally {
-      setIsUpdating(false)
-    }
-  }
-
-  if (!user) return null
-
-  const getInitials = (name: string) => {
-    return name
-      .split(" ")
-      .map((n) => n[0])
-      .join("")
-      .toUpperCase()
-      .slice(0, 2)
-  }
-
-  const completedEnrollments = enrollments.filter((e) => e.completed_at)
-  const avgProgress =
-    enrollments.length > 0 ? enrollments.reduce((sum, e) => sum + e.progress_percentage, 0) / enrollments.length : 0
-
-  const avgQuizScore =
-    quizAttempts.length > 0
-      ? quizAttempts.reduce((sum, a) => sum + (a.score / a.max_score) * 100, 0) / quizAttempts.length
-      : 0
-
-  const getStatusBadge = (status: User["status"]) => {
-    switch (status) {
-      case "active":
-        return <Badge className="bg-green-500/10 text-green-500">Active</Badge>
-      case "locked":
-        return <Badge className="bg-orange-500/10 text-orange-500">Locked</Badge>
-      case "suspended":
-        return <Badge className="bg-yellow-500/10 text-yellow-500">Suspended</Badge>
-      case "banned":
-        return <Badge variant="destructive">Banned</Badge>
-      case "disabled":
-        return <Badge className="bg-gray-500/10 text-gray-500">Disabled</Badge>
-      case "deleted":
-        return <Badge variant="destructive">Deleted</Badge>
-      default:
-        return <Badge variant="outline">{status}</Badge>
-    }
-  }
-
-  const getRoleBadge = (role: User["role"]) => {
-    switch (role) {
-      case "admin":
-        return <Badge className="bg-purple-500/10 text-purple-500">Admin</Badge>
-      case "moderator":
-        return <Badge className="bg-blue-500/10 text-blue-500">Moderator</Badge>
-      case "instructor":
-        return <Badge className="bg-cyan-500/10 text-cyan-500">Instructor</Badge>
-      default:
-        return <Badge variant="outline">User</Badge>
-    }
-  }
-
-  return (
-    <>
-      <Dialog open={open} onOpenChange={onClose}>
-        <DialogContent className="max-w-5xl h-[90vh] flex flex-col">
-          <DialogHeader>
-            <DialogTitle>User Details</DialogTitle>
-          </DialogHeader>
-
-          <div className="flex-1 overflow-y-auto">
-            <div className="space-y-6">
-              {/* User Header */}
-              <div className="flex items-start gap-4">
-                <Avatar className="h-20 w-20">
-                  <AvatarImage src={user.avatar_url || "/placeholder.svg"} />
-                  <AvatarFallback className="text-2xl">{getInitials(user.full_name)}</AvatarFallback>
-                </Avatar>
-
-                <div className="flex-1">
-                  <div className="flex items-center gap-2 mb-1">
-                    <h2 className="text-2xl font-bold">{user.full_name}</h2>
-                    {user.email_verified ? (
-                      <MailCheck className="h-5 w-5 text-green-500" />
-                    ) : (
-                      <Mail className="h-5 w-5 text-muted-foreground" />
-                    )}
-                  </div>
-                  <p className="text-muted-foreground">{user.email}</p>
-                  <p className="text-sm text-muted-foreground">@{user.username}</p>
-
-                  <div className="flex gap-2 mt-3">
-                    {getStatusBadge(user.status)}
-                    {getRoleBadge(user.role)}
-                    {user.email_verified ? (
-                      <Badge variant="outline">Email Verified</Badge>
-                    ) : (
-                      <Badge variant="secondary">Email Unverified</Badge>
-                    )}
-                  </div>
-                </div>
-              </div>
-
-              {/* Admin Actions */}
-              <Card className="p-4">
-                <h3 className="font-semibold mb-4 flex items-center gap-2">
-                  <Shield className="h-4 w-4" />
-                  Admin Actions
-                </h3>
-                <div className="grid grid-cols-2 gap-4">
-                  <div>
-                    <Label htmlFor="role">Change Role</Label>
-                    <Select value={user.role} onValueChange={handleRoleChange} disabled={isUpdating}>
-                      <SelectTrigger id="role">
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="user">User</SelectItem>
-                        <SelectItem value="instructor">Instructor</SelectItem>
-                        <SelectItem value="moderator">Moderator</SelectItem>
-                        <SelectItem value="admin">Admin</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div>
-
-                  <div>
-                    <Label htmlFor="tenant">Tenant</Label>
-                    <Input
-                      id="tenant"
-                      value={user.tenant}
-                      onChange={(e) => handleTenantChange(e.target.value)}
-                      disabled={isUpdating}
-                      placeholder="Enter tenant name"
-                    />
-                  </div>
-                </div>
-
-                <div className="flex gap-2 mt-4">
-                  {user.status === "locked" ? (
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      onClick={handleUnlockAccount}
-                      disabled={isUpdating}
-                      className="gap-2 bg-transparent"
-                    >
-                      <Unlock className="h-4 w-4" />
-                      Unlock Account
-                    </Button>
-                  ) : (
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      onClick={() => setShowLockDialog(true)}
-                      disabled={isUpdating || user.status === "deleted"}
-                      className="gap-2"
-                    >
-                      <Lock className="h-4 w-4" />
-                      Lock Account
-                    </Button>
-                  )}
-
-                  {user.status === "deleted" ? (
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      onClick={handleRestore}
-                      disabled={isUpdating}
-                      className="gap-2 bg-transparent"
-                    >
-                      <RotateCcw className="h-4 w-4" />
-                      Restore Account
-                    </Button>
-                  ) : (
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      onClick={() => setShowDeleteDialog(true)}
-                      disabled={isUpdating}
-                      className="gap-2"
-                    >
-                      <Trash2 className="h-4 w-4" />
-                      Soft Delete
-                    </Button>
-                  )}
-
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    onClick={() => setShowResetPasswordDialog(true)}
-                    disabled={isUpdating || user.status === "deleted"}
-                    className="gap-2"
-                  >
-                    <KeyRound className="h-4 w-4" />
-                    Reset Password
-                  </Button>
-                </div>
-              </Card>
-
-              {/* Stats Grid */}
-              <div className="grid grid-cols-4 gap-4">
-                <Card className="p-4">
-                  <div className="flex items-center gap-2 text-muted-foreground mb-1">
-                    <Trophy className="h-4 w-4" />
-                    <span className="text-sm">Total Points</span>
-                  </div>
-                  <p className="text-2xl font-bold">{user.total_points.toLocaleString()}</p>
-                </Card>
-
-                <Card className="p-4">
-                  <div className="flex items-center gap-2 text-muted-foreground mb-1">
-                    <TrendingUp className="h-4 w-4" />
-                    <span className="text-sm">Current Streak</span>
-                  </div>
-                  <p className="text-2xl font-bold">{user.current_streak} days</p>
-                </Card>
-
-                <Card className="p-4">
-                  <div className="flex items-center gap-2 text-muted-foreground mb-1">
-                    <BookOpen className="h-4 w-4" />
-                    <span className="text-sm">Enrollments</span>
-                  </div>
-                  <p className="text-2xl font-bold">{enrollments.length}</p>
-                  <p className="text-xs text-muted-foreground">{completedEnrollments.length} completed</p>
-                </Card>
-
-                <Card className="p-4">
-                  <div className="flex items-center gap-2 text-muted-foreground mb-1">
-                    <HelpCircle className="h-4 w-4" />
-                    <span className="text-sm">Quiz Attempts</span>
-                  </div>
-                  <p className="text-2xl font-bold">{quizAttempts.length}</p>
-                  <p className="text-xs text-muted-foreground">{avgQuizScore.toFixed(0)}% avg score</p>
-                </Card>
-              </div>
-
-              {/* Account Information */}
-              <Card className="p-4">
-                <h3 className="font-semibold mb-4">Account Information</h3>
-                <div className="grid grid-cols-3 gap-4 text-sm">
-                  <div>
-                    <span className="text-muted-foreground flex items-center gap-2 mb-1">
-                      <Mail className="h-3 w-3" />
-                      User ID
-                    </span>
-                    <p className="font-mono text-xs break-all">{user.id}</p>
-                  </div>
-                  <div>
-                    <span className="text-muted-foreground flex items-center gap-2 mb-1">
-                      <Mail className="h-3 w-3" />
-                      Email
-                    </span>
-                    <p className="font-medium">{user.email}</p>
-                  </div>
-                  <div>
-                    <span className="text-muted-foreground flex items-center gap-2 mb-1">
-                      <Shield className="h-3 w-3" />
-                      Role
-                    </span>
-                    <p className="font-medium capitalize">{user.role}</p>
-                  </div>
-                  <div>
-                    <span className="text-muted-foreground flex items-center gap-2 mb-1">
-                      <Building2 className="h-3 w-3" />
-                      Tenant
-                    </span>
-                    <p className="font-medium">{user.tenant}</p>
-                  </div>
-                  <div>
-                    <span className="text-muted-foreground flex items-center gap-2 mb-1">
-                      <Calendar className="h-3 w-3" />
-                      Created
-                    </span>
-                    <p className="font-medium">{new Date(user.created_at).toLocaleDateString()}</p>
-                  </div>
-                  <div>
-                    <span className="text-muted-foreground flex items-center gap-2 mb-1">
-                      <Calendar className="h-3 w-3" />
-                      Updated
-                    </span>
-                    <p className="font-medium">{new Date(user.updated_at).toLocaleDateString()}</p>
-                  </div>
-                  <div>
-                    <span className="text-muted-foreground flex items-center gap-2 mb-1">
-                      <Globe className="h-3 w-3" />
-                      Last Login
-                    </span>
-                    <p className="font-medium">
-                      {user.last_login_at ? new Date(user.last_login_at).toLocaleString() : "Never"}
-                    </p>
-                  </div>
-                  <div>
-                    <span className="text-muted-foreground flex items-center gap-2 mb-1">
-                      <Globe className="h-3 w-3" />
-                      Last Login IP
-                    </span>
-                    <p className="font-mono text-xs">{user.last_login_ip || "N/A"}</p>
-                  </div>
-                  {user.lockout_until && (
-                    <div>
-                      <span className="text-muted-foreground flex items-center gap-2 mb-1">
-                        <Lock className="h-3 w-3" />
-                        Locked Until
-                      </span>
-                      <p className="font-medium text-orange-500">{new Date(user.lockout_until).toLocaleDateString()}</p>
-                    </div>
-                  )}
-                  {user.deleted_at && (
-                    <div>
-                      <span className="text-muted-foreground flex items-center gap-2 mb-1">
-                        <Trash2 className="h-3 w-3" />
-                        Deleted At
-                      </span>
-                      <p className="font-medium text-red-500">{new Date(user.deleted_at).toLocaleDateString()}</p>
-                    </div>
-                  )}
-                </div>
-              </Card>
-
-              {/* Activity Tabs */}
-              <Tabs defaultValue="enrollments">
-                <TabsList className="grid w-full grid-cols-2">
-                  <TabsTrigger value="enrollments">Enrollments ({enrollments.length})</TabsTrigger>
-                  <TabsTrigger value="quizzes">Quiz Attempts ({quizAttempts.length})</TabsTrigger>
-                </TabsList>
-
-                <TabsContent value="enrollments" className="space-y-3 mt-4">
-                  {enrollments.length === 0 ? (
-                    <p className="text-center text-muted-foreground py-8">No enrollments yet</p>
-                  ) : (
-                    enrollments.map((enrollment) => (
-                      <Card key={enrollment.id} className="p-4">
-                        <div className="flex items-center justify-between">
-                          <div>
-                            <p className="font-medium">Lesson #{enrollment.lesson_id}</p>
-                            <p className="text-sm text-muted-foreground">
-                              Enrolled: {new Date(enrollment.enrolled_at).toLocaleDateString()}
-                            </p>
-                          </div>
-                          <div className="text-right">
-                            <p className="text-sm font-medium">{enrollment.progress_percentage}% complete</p>
-                            {enrollment.completed_at && (
-                              <Badge variant="outline" className="mt-1">
-                                Completed
-                              </Badge>
-                            )}
-                          </div>
-                        </div>
-                      </Card>
-                    ))
-                  )}
-                </TabsContent>
-
-                <TabsContent value="quizzes" className="space-y-3 mt-4">
-                  {quizAttempts.length === 0 ? (
-                    <p className="text-center text-muted-foreground py-8">No quiz attempts yet</p>
-                  ) : (
-                    quizAttempts.map((attempt) => (
-                      <Card key={attempt.id} className="p-4">
-                        <div className="flex items-center justify-between">
-                          <div>
-                            <p className="font-medium">Quiz #{attempt.quiz_id}</p>
-                            <p className="text-sm text-muted-foreground">
-                              Started: {new Date(attempt.started_at).toLocaleDateString()}
-                            </p>
-                          </div>
-                          <div className="text-right">
-                            <p className="text-lg font-bold">
-                              {attempt.score}/{attempt.max_score}
-                            </p>
-                            <p className="text-sm text-muted-foreground">
-                              {((attempt.score / attempt.max_score) * 100).toFixed(0)}%
-                            </p>
-                            {attempt.completed_at && (
-                              <Badge variant="outline" className="mt-1">
-                                Completed
-                              </Badge>
-                            )}
-                          </div>
-                        </div>
-                      </Card>
-                    ))
-                  )}
-                </TabsContent>
-              </Tabs>
-            </div>
-          </div>
-        </DialogContent>
-      </Dialog>
-
-      {/* Lock Account Dialog */}
-      <AlertDialog open={showLockDialog} onOpenChange={setShowLockDialog}>
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Lock User Account</AlertDialogTitle>
-            <AlertDialogDescription>
-              This will prevent the user from accessing their account until the lockout period expires.
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <div className="py-4">
-            <Label htmlFor="lockout-days">Lockout Duration (days)</Label>
-            <Input
-              id="lockout-days"
-              type="number"
-              value={lockoutDays}
-              onChange={(e) => setLockoutDays(e.target.value)}
-              min="1"
-              max="365"
-              className="mt-2"
-            />
-          </div>
-          <AlertDialogFooter>
-            <AlertDialogCancel>Cancel</AlertDialogCancel>
-            <AlertDialogAction onClick={handleLockAccount} disabled={isUpdating}>
-              Lock Account
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
-
-      {/* Soft Delete Dialog */}
-      <AlertDialog open={showDeleteDialog} onOpenChange={setShowDeleteDialog}>
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Soft Delete User Account</AlertDialogTitle>
-            <AlertDialogDescription>
-              This will mark the user account as deleted. The account can be restored later if needed.
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel>Cancel</AlertDialogCancel>
-            <AlertDialogAction onClick={handleSoftDelete} disabled={isUpdating} className="bg-destructive">
-              Delete Account
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
-
-      {/* Reset Password Dialog */}
-      <AlertDialog open={showResetPasswordDialog} onOpenChange={setShowResetPasswordDialog}>
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Reset User Password</AlertDialogTitle>
-            <AlertDialogDescription>
-              This will send a password reset email to {user?.email}. The user will need to follow the link to set a new
-              password.
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel>Cancel</AlertDialogCancel>
-            <AlertDialogAction onClick={handleResetPassword} disabled={isUpdating}>
-              Send Reset Email
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
-    </>
-  )
-}
